% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergenceModelR-package.R
\docType{package}
\name{emergenceModelR-package}
\alias{emergenceModelR}
\alias{emergenceModelR-package}
\title{emergenceModelR: Educational Simulations for Emergence and Complexity}
\description{
\code{emergenceModelR} provides simplified computational simulations for teaching
and exploring emergence, self-organization, complexity, cellular automata,
agent interactions, network growth, and information-based metrics.
}
\details{
These functions are educational abstractions. They do not fully model real
biological, social, ecological, cognitive, or conscious systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NoushinN/emergenceModelR}
  \item \url{https://noushinn.github.io/emergenceModelR/}
  \item Report bugs at \url{https://github.com/NoushinN/emergenceModelR/issues}
}

}
\author{
\strong{Maintainer}: Noushin Nabavi \email{NoushinN@users.noreply.github.com}

Authors:
\itemize{
  \item Noushin Nabavi \email{NoushinN@users.noreply.github.com}
}

}
\keyword{internal}
